+++ rpc/rpc.go
@@ -177,38 +177,47 @@ func (l *Link) RPC(ctx context.Context, msg *amqp.Message) (*Response, error) {
 }
 
 // Close the link receiver, sender and session
-func (l *Link) Close() error {
-	if err := l.closeReceiver(); err != nil {
-		_ = l.closeSender()
-		_ = l.closeSession()
+func (l *Link) Close(ctx context.Context) error {
+	span, ctx := tracing.StartSpanFromContext(ctx, "az-amqp-common.rpc.Close")
+	span.Finish()
+
+	if err := l.closeReceiver(ctx); err != nil {
+		_ = l.closeSender(ctx)
+		_ = l.closeSession(ctx)
 		return err
 	}
 
-	if err := l.closeSender(); err != nil {
-		_ = l.closeSession()
+	if err := l.closeSender(ctx); err != nil {
+		_ = l.closeSession(ctx)
 		return err
 	}
 
-	return l.closeSession()
+	return l.closeSession(ctx)
 }
 
-func (l *Link) closeReceiver() error {
+func (l *Link) closeReceiver(ctx context.Context) error {
+	span, ctx := tracing.StartSpanFromContext(ctx, "az-amqp-common.rpc.closeReceiver")
+	span.Finish()
 	if l.receiver != nil {
-		return l.receiver.Close()
+		return l.receiver.Close(ctx)
 	}
 	return nil
 }
 
-func (l *Link) closeSender() error {
+func (l *Link) closeSender(ctx context.Context) error {
+	span, ctx := tracing.StartSpanFromContext(ctx, "az-amqp-common.rpc.closeSender")
+	span.Finish()
 	if l.sender != nil {
-		return l.sender.Close()
+		return l.sender.Close(ctx)
 	}
 	return nil
 }
 
-func (l *Link) closeSession() error {
+func (l *Link) closeSession(ctx context.Context) error {
+	span, ctx := tracing.StartSpanFromContext(ctx, "az-amqp-common.rpc.closeSession")
+	span.Finish()
 	if l.session != nil {
-		return l.session.Close()
+		return l.session.Close(ctx)
 	}
 	return nil
 }